refactor schema; include [new] `$inferInsert` type
