+++ src/v2/db/schema/user/userConnections.ts
@@ -1,0 +1,46 @@
+import { tableNames } from "@/v2/db/drizzle"
+import { relations } from "drizzle-orm"
+import { sqliteTable, text, index } from "drizzle-orm/sqlite-core"
+import { users } from "./user"
+
+/*
+NOTE: This file will be expanded on in the future, but for now it's just for Discord.
+- This is used to link an account to ID but also can be set by initial Discord OAuth.
+*/
+
+export const socialsConnection = sqliteTable(
+    tableNames.socialsConnection,
+    {
+        id: text("id").primaryKey(),
+        userId: text("user_id")
+            .notNull()
+            .references(() => users.id, {
+                onUpdate: "cascade",
+                onDelete: "cascade",
+            }),
+        discordId: text("discord_id"),
+    },
+    (socialsConnection) => {
+        return {
+            userIdx: index("socials_connection_user_id_idx").on(
+                socialsConnection.userId
+            ),
+            discordIdIdx: index("socials_connection_discord_id_idx").on(
+                socialsConnection.discordId
+            ),
+        }
+    }
+)
+
+export type SocialsConnection = typeof socialsConnection.$inferSelect
+export type NewSocialsConnection = typeof socialsConnection.$inferInsert
+
+export const socialsConnectionRelations = relations(
+    socialsConnection,
+    ({ one }) => ({
+        user: one(users, {
+            fields: [socialsConnection.userId],
+            references: [users.id],
+        }),
+    })
+)
